+++ example/get_orders/test/get_orders_test.dart
@@ -1,0 +1,8 @@
+import 'package:get_orders/get_orders.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('calculate', () {
+    expect(calculate(), 42);
+  });
+}